next pages to app init
