+++ package.json
@@ -9,23 +9,21 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@next/font": "13.1.6",
-    "@types/node": "18.11.19",
-    "@types/react": "18.0.27",
-    "@types/react-dom": "18.0.10",
-    "eslint": "8.33.0",
-    "eslint-config-next": "13.1.6",
-    "next": "13.1.6",
+    "@types/node": "18.15.12",
+    "@types/react": "^18.0.37",
+    "@types/react-dom": "18",
+    "eslint": "8.38.0",
+    "eslint-config-next": "^13.3.1-canary.4",
+    "next": "^13.3.1-canary.4",
     "next-themes": "^0.2.1",
     "qs": "^6.11.1",
     "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "typescript": "4.9.5"
+    "react-dom": "18.2.0"
   },
   "devDependencies": {
-    "@tailwindcss/line-clamp": "^0.4.2",
     "autoprefixer": "^10.4.13",
     "postcss": "^8.4.21",
-    "tailwindcss": "^3.2.4"
+    "tailwindcss": "^3.2.4",
+    "typescript": "5.1.0-beta"
   }
 }
