+++ src/app/globals.css
@@ -1,0 +1,31 @@
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
+
+html{
+  @apply scroll-smooth font-sans;
+}
+body {
+  overflow-x: hidden;
+  @apply bg-white dark:bg-darkBg overflow-hidden overflow-y-auto transition
+}
+
+.clipTriangle{
+  -webkit-clip-path: polygon(50% 0, 0 50%, 50% 97%);
+clip-path: polygon(50% 0, 0 50%, 50% 97%);
+}
+
+.clipTriangleTopLeft{
+  -webkit-clip-path: polygon(0 0, 0 100%, 100% 0);
+  clip-path: polygon(0 0, 0 100%, 100% 0);
+}
+
+.clipTriangleBottomRight{
+  -webkit-clip-path: polygon(0 100%, 100% 100%, 100% 0);
+  clip-path: polygon(0 100%, 100% 100%, 100% 0);
+}
+
+.clipTriangleTopRight{
+  -webkit-clip-path: polygon(0 0, 100% 100%, 100% 0);
+  clip-path: polygon(0 0, 100% 100%, 100% 0);
+}
