+++ src/components/elements/ThemeSwitcher.tsx
@@ -1,4 +1,4 @@
-
+'use client'
 import { useTheme } from "next-themes"
 
 function ThemeSwitcher() {
