+++ types/jsx.d.ts
@@ -1,0 +1,11 @@
+namespace JSX {
+	type FunctionComponent = (props: any) => Element | null
+	type FunctionComponentWithPromiseReturn = (
+		props: any
+	) => Promise<Element | null>
+
+	export type ElementType =
+		| keyof IntrinsicElements
+		| FunctionComponent
+		| FunctionComponentWithPromiseReturn
+}
